Fix multiple maps rendering when changing routes
Every time that we load in the data for a different year, the map
won't rerender with the new data. Instead, it will just append an entirely
new svg map with the new data. This commit solves it by removing the
existing svg element and then rendering a new one.