Fix broken dependency for production build
It appears that the current version of ember-cli-d3 breaks
in production; it works perfectly in development. That Ember plugin
attempts to access Ember's testing tools which are not built into the
production distributable which causes the app to crash on load. This
commit fixes that by removing the faulty plugin and packages the actual
d3 npm module into the app using ember-browserify.